Creating Jonno Profile Draft
Seriously limited info on Jonno, wasn't really involved outside of active shenanigans, romance with Telia (and occasional humorously redacted descriptions thereof), and sudden departure to Zipangu.  On the other hand, unless the original author gets involved, this guy and the others of his group have a lot of room for detail addons as needed.